# apps/api-java/src/main/java/org/opencivic/signalos/web/dto/CreateCommunityThreadMessageRequest.java
package org.opencivic.signalos.web.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import java.util.UUID;

public record CreateCommunityThreadMessageRequest(
    @NotNull UUID sourceCommunityId,
    @NotBlank String content,
    UUID parentMessageId
) {}