# apps/api-java/src/main/java/org/opencivic/signalos/web/dto/UpdateCommunityRoleRequest.java
package org.opencivic.signalos.web.dto;

import jakarta.validation.constraints.NotNull;
import org.opencivic.signalos.domain.CommunityRole;

public record UpdateCommunityRoleRequest(
    @NotNull CommunityRole role
) {}